Improving file split to avoid slow tests
